refactor(config): drop unused explorer headers
